feat: refactored initial code and files structure
