+++ README.md
@@ -1,3 +1,8 @@
 # SODV1201_Project_SharedWorkspace
 Course Project of SODV1201 (Introduction to Web Programming)
 
+# Instalation
+run `npm install` to install **node_modules** from **package.json**
+
+# Run Code
+run `node server.js` to run the node server or run `nodemon server.js` to automatically restarting the node application when file changes in the directory are detected.
