+++ package.json
@@ -1,11 +1,12 @@
 {
-  "name": "assignment-2",
+  "name": "SODV1201_Project_SharedWorkspace",
   "version": "1.0.0",
   "description": "",
   "main": "server.js",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
+    "start": "node server.js",
+    "dev": "nodemon server.js" 
   },
   "keywords": [],
   "author": "",
