+++ public/homePageOwner.html
@@ -6,7 +6,7 @@
     <title>Shared Workspace WebApp</title>
 </head>
 <body>
-    <h2>Home Page Owner</h2>
+    <h1>Homepage Owner</h1>
     
     <form action="/userRegistration" method="post">
         <label for="userName">Username:</label>
@@ -16,6 +16,8 @@
         <input type="password" id="password" name="password" required>
         <input type="submit" value="Register">
       </form>
+
+      
       <script>
         function fillFee(){
             const status = document.getElementById("status")
