+++ public/index.html
@@ -6,7 +6,7 @@
     <title>Shared Workspace WebApp</title>
 </head>
 <body>
-    <h2>Login</h2>
+    <h1>Login Page</h1>
     
     <form action="/" method="post">
         <label for="userName">Username:</label>
@@ -17,6 +17,8 @@
         <input type="submit" value="Register">
         <a id="userRegistrationButton" href="./userRegistration.html">New User?</a>
       </form>
+
+      
       <script>
         function fillFee(){
             const status = document.getElementById("status")
