+++ public/propertyRegistration.html
@@ -6,7 +6,7 @@
     <title>Shared Workspace WebApp</title>
 </head>
 <body>
-    <h2>BVC Sports Club</h2>
+    <h1>Property Registration Page</h1>
     
     <form action="/" method="post">
         <label for="fullName">Full name:</label>
@@ -34,6 +34,8 @@
         <br>
         <input type="submit" value="Register">
       </form>
+
+      
       <script>
         function fillFee(){
             const status = document.getElementById("status")
