+++ public/userRegistration.html
@@ -6,7 +6,7 @@
     <title>Shared Workspace WebApp</title>
 </head>
 <body>
-    <h2>Registration</h2>
+    <h1>User Registration Page</h1>
     
     <form action="/userRegistration" method="post">
         <label for="userName">Username:</label>
