--- users.json
@@ -1,1 +1,0 @@
-[{"userName":"Matheus","password":"12345","status":"owner"},{"userName":"Daniel","password":"54321","status":"coworker"}]
